refactor(admin): extract health check handler module
